+++ _includes/changelog.md
@@ -5,6 +5,11 @@
 - Update label positioning
 - Update class system to remove implicit initialize and transform cascade
 - Remove deprecated xy extension and Object-style layout
+- __0.15.1__ Fix Overlay.position and labels with stacked bars bugs
+- __0.15.2__ Fix Gridlines onExit bug and slide down Bars on exit
+- __0.15.3__ Fix layout calculation and axis transition bugs
+- __0.15.4__ Fix component centering and mousemove in IE bugs
+- __0.15.5__ Fix __proto__ and Overlay transform issues
 
 # 0.14.0